# client/src/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bookify API",
    "description": "API del back-end di Bookify",
    "version": "v1"
  },
  "paths": {
    "/api/Accommodation": {
      "get": {
        "tags": [
          "Accommodation"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isDescending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "includes",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDtoPagedResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Accommodation"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Address.City",
                  "Address.Country",
                  "Address.Number",
                  "Address.PostalCode",
                  "Address.Province",
                  "Address.Street",
                  "ImgFile",
                  "Name",
                  "Type"
                ],
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Type": {
                    "$ref": "#/components/schemas/AccommodationType"
                  },
                  "ImgFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "OwnerId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Address.Street": {
                    "type": "string"
                  },
                  "Address.Number": {
                    "type": "string"
                  },
                  "Address.City": {
                    "type": "string"
                  },
                  "Address.Province": {
                    "type": "string"
                  },
                  "Address.PostalCode": {
                    "type": "string"
                  },
                  "Address.Country": {
                    "type": "string"
                  },
                  "Address.AdditionalInfo": {
                    "type": "string"
                  },
                  "Address.Latitude": {
                    "type": "number",
                    "format": "double"
                  },
                  "Address.Longitude": {
                    "type": "number",
                    "format": "double"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Type": {
                  "style": "form"
                },
                "ImgFile": {
                  "style": "form"
                },
                "OwnerId": {
                  "style": "form"
                },
                "Address.Street": {
                  "style": "form"
                },
                "Address.Number": {
                  "style": "form"
                },
                "Address.City": {
                  "style": "form"
                },
                "Address.Province": {
                  "style": "form"
                },
                "Address.PostalCode": {
                  "style": "form"
                },
                "Address.Country": {
                  "style": "form"
                },
                "Address.AdditionalInfo": {
                  "style": "form"
                },
                "Address.Latitude": {
                  "style": "form"
                },
                "Address.Longitude": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Accommodation/{id}": {
      "get": {
        "tags": [
          "Accommodation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "includes",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Accommodation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccommodationUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AccommodationUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AccommodationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Accommodation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Accommodation": {
        "required": [
          "imgUrl",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/AccommodationType"
          },
          "imgUrl": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "integer",
            "format": "int32"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "addressId": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Booking"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AccommodationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "imgUrl": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/AccommodationType"
          },
          "ownerId": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Booking"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AccommodationDtoPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccommodationDto"
            },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "hasPrevious": {
            "type": "boolean",
            "readOnly": true
          },
          "hasNext": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "AccommodationType": {
        "enum": [
          "Hotel",
          "Apartment",
          "Resort",
          "Villa",
          "Hostel"
        ],
        "type": "string"
      },
      "AccommodationUpdate": {
        "required": [
          "address",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/AccommodationType"
          },
          "imgFile": {
            "type": "string",
            "format": "binary",
            "nullable": true
          },
          "ownerId": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          }
        },
        "additionalProperties": false
      },
...
      "Address": {
        "required": [
          "city",
          "country",
          "number",
          "postalCode",
          "province",
          "street"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "province": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "additionalInfo": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddressDto": {
        "required": [
          "city",
          "country",
          "number",
          "postalCode",
          "province",
          "street"
        ],
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "province": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "additionalInfo": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Booking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "checkInDate": {
            "type": "string",
            "format": "date-time"
          },
          "checkOutDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "accommodationId": {
            "type": "integer",
            "format": "int32"
          },
          "accommodation": {
            "$ref": "#/components/schemas/Accommodation"
          }
        },
        "additionalProperties": false
      },
...
      "BookingStatus": {
        "enum": [
          "Pending",
          "Confirmed",
          "Completed",
          "Cancelled"
        ],
        "type": "string"
      },
...
      "User": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "phoneNumber",
          "username",
          "uuid"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "registrationDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "addressId": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "authorizationUrl": "http://localhost:8080/realms/bookify/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:8080/realms/bookify/protocol/openid-connect/token",
            "scopes": {
              "openid": "OpenID Connect scope",
              "profile": "Profile information",
              "email": "Email information"
            }
          },
          "authorizationCode": {
            "authorizationUrl": "http://localhost:8080/realms/bookify/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:8080/realms/bookify/protocol/openid-connect/token",
            "scopes": {
              "openid": "OpenID Connect scope",
              "profile": "Profile information",
              "email": "Email information"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "openid",
        "profile",
        "email"
      ]
    }
  ]
}